Some how I never pushed the changes on the laptop... oops

+++ logging_receiver_dsd.cc
@@ -48,7 +48,8 @@ log_dsd::log_dsd(float f, float c, long t, int n)
 	float offset = center - (f*1000000);
 
 	int samp_per_sym = 10;
-	double samp_rate = 5000000;	int decim = 80;
+	double samp_rate = 5000000;	
+	double decim = 80;
 	float xlate_bandwidth = 14000; //24260.0;
 	float channel_rate = 4800 * samp_per_sym;
 	double pre_channel_rate = double(samp_rate/decim);
